ocpiriun utility (run app from xml), array value string bug fixed, endpoint-per-port in hdl containers -- major debug/improvements to the ocpihdl utility for etherhet data plane debug -- data transfer endpoint enhancements to share more code Makefile.ocpi.common: easier way to preload drivers via MAIN_DRIVERS, but still mangled -- include transfer drivers library (not obj) on DATA_TRANSFER_LIBS -- define ALL_CONTAINER_LIBS Makefile.ocpi.for-pkg: enabling simpler main programs without a separate bin directory Makefile.ocpi.for-src: ditto Makefile.ocpi.sub-bin: ditto (new file) components/dsample_complex.rcc/dsample_complex.c: removed - was checked in erroneously components/file_read.rcc/file_read.c: fix temporary breakage that broke examples components/file_read.rcc/file_read.xml: remote protocol core/application/Makefile.ocpi.imports: add libraries and drivers for ocpirun core/application/include/OcpiApplication.h: arg name change policy->params core/application/include/OcpiApplicationApi.h: ditto core/application/src/OcpiApplication.cxx: delegate more policy parsing to OU::Assembly -- allow policy overrides as parameters to the application constructor core/container/hdl_container/bin/src/ocpihdl_main.cxx: ensure correct discovery mode per command -- better printing in emulator -- allow worker settings in emulator, not just admin -- many receiveRDMA improvements core/container/hdl_container/impl/include/EtherDriver.h: ensure correct discovery mode with explicit arg core/container/hdl_container/impl/include/HdlDriver.h: ditto core/container/hdl_container/impl/src/EtherDriver.cxx:  ditto -- plus tighten up command packet lengths so wireshark is happier core/container/hdl_container/impl/include/HdlAccess.h: initialize device with protocol core/container/hdl_container/impl/src/HdlAccess.cxx: ditto core/container/hdl_container/impl/include/HdlContainer.h: remove endpoint from container - device has it core/container/hdl_container/impl/src/HdlContainer.cxx: more dynamic endpoint creation (multiple per device) -- remote endpoint is now per port -- changes for generic external ports core/container/hdl_container/impl/src/HdlDriver.cxx: discovery mode args core/container/hdl_container/impl/src/PciDriver.cxx: ditto, plus deferred endpoint creation (now per port) core/container/interfaces/api/include/OcpiContainerPort.h: improve getPreferredProtocol interface core/container/interfaces/api/src/OcpiContainerPort.cxx: ditto core/container/ocl_container/impl/src/OcpiOclContainer.cxx: export destructor as public core/dataplane/rdma_driver_interface/interfaces/include/DtHandshakeControl.h: remove some OS:: integer types core/dataplane/rdma_driver_interface/interfaces/include/DtIntEventHandler.h: ditto core/dataplane/rdma_driver_interface/interfaces/include/DtSharedMemoryInterface.h: ditto + 64->32 bit offsets -- plus add the "address" field to the endpoint structure since it is fairly generic -- i.e. it is available for all endpoints that need it, so more code can be generic (not all). core/dataplane/rdma_driver_interface/interfaces/include/DtSharedMemoryInternal.h: ditto core/dataplane/rdma_driver_interface/interfaces/include/DtTransferInternal.h: ditto core/dataplane/rdma_driver_interface/interfaces/include/OcpiBaseSmemServices.h: ditto core/dataplane/rdma_driver_interface/interfaces/include/OcpiRDTInterface.h: ditto core/dataplane/rdma_driver_interface/interfaces/src/DtIntEventHandler.cxx: ditto core/dataplane/rdma_driver_interface/interfaces/src/DtSharedMemoryInternal.cxx: ditto -- plus handling of new address field -- plus cleanup bug fix when factory isn't set yet core/dataplane/rdma_driver_interface/interfaces/src/DtTransferInternal.cxx: ditto core/dataplane/rdma_drivers/datagram/include/DtDataGramXfer.h: ditto core/dataplane/rdma_drivers/datagram/src/DtDataGramXfer.cxx: reinstate asserts core/dataplane/rdma_drivers/ofed/src/DtOfedXfer.cxx: remove OS:: int types core/dataplane/rdma_drivers/pci/src/DtPCIPioXfer.cxx: ditto -- plus change delimiter in endpoint after address to ';' core/dataplane/rdma_drivers/sockets/src/DtSocketsXfer.cxx: ditto core/dataplane/rdma_drivers/udp/src/DtUDPSocketXfer.cxx: ditto core/dataplane/rdma_smb/host/src/OcpiHostSmemServices.cxx: ditto, cleanup core/dataplane/transport/impl/include/OcpiPort.h: ditto core/dataplane/transport/impl/include/OcpiTransport.h: remove unused createPort -- add utility to fill descriptors from endpoints core/dataplane/transport/impl/src/OcpiTransport.cxx: ditto core/dataplane/transport/impl/src/OcpiPort.cxx: more OS:: intt types and 64->32 offsets core/library/src/OcpiLibraryAssembly.cxx: better protocol mismatch error message core/local/util/assembly/include/OcpiUtilAssembly.h: move policy parsing here from app core/local/util/assembly/src/OcpiUtilAssembly.cxx: move policy parsing here from app -- allow policy setting via simple attributes core/local/util/driver/include/OcpiDriverManager.h: add generic option to suppress discovery for a driver - some utilities need to suppress discovery for certain managers core/local/util/driver/src/OcpiDriverManager.cxx: ditto, plus more debug logging core/local/util/property/include/OcpiUtilValue.h: bug fix to allow simple single dimension array values -- do not require braces for single-dimension array values -- support unparse in both non-append and append mode core/local/util/property/src/OcpiUtilValue.cxx: ditto -- plus two nasty bug fixes in Value constructor and printer that broke all the tests core/local/util/property/src/OcpiUtilDataTypes.cxx: better error messsages -- plus redundant "default" attribute in xml printing removed. core/local/util/property/src/OcpiUtilProtocol.cxx: better test output printing core/local/util/res/include/OcpiRes.h: 32 bit sizes (from 64) core/local/util/res/src/OcpiAlloc.cxx: ditto tools/cdk/examples/bias2_xml/Makefile: add make target for running ocpirun on xml tools/cdk/examples/bias_xml/Makefile: ditto tools/cdk/include/hdl/hdl-pre.mk: bug fix to allow unaltered skeleton removal tools/cdk/include/hdl/hdl-worker.mk: early work to reinstate VHDL support tools/cdk/include/hdl/icarus.mk: turn on warnings tools/cdk/include/xxx-worker.mk: support running make in the worker's impl directory -- and cleaning/removal of unaltered skeletons tools/cdk/ocpidds/bin/src/ocpidds_main.cxx: add another test mode to use xml proto file for tests tools/cdk/ocpidds/ocpidds/include/ocpidds.h: ditto tools/cdk/ocpidds/ocpidds/src/dataTypeTest.cxx: ditto tools/cdk/ocpigen/ocpigen/src/wip_HDL.cxx: some more VHDL code generation reorganization components/file_write.rcc/file_write.c: remove printf components/specs/bias_spec.xml: use new 32 bit protocol core/dataplane/rdma_drivers/udp/src/DtUDPSocketXfer.cxx: OS int types, 64->32 offsets components/specs/stream32_protocol.xml: new file for 32 bit protocol for bias core/application/src/ocpirun_main.cxx: new utility to run app from xml with no C++ tools/cdk/examples/bias2_xml/bias2.xml: new pure xml file for bias2 example tools/cdk/examples/bias_xml/bias.xml: ditto for bias tools/cdk/export/bin/darwin-x86_64/ocpirun: symlink for ocpirun tools/cdk/export/bin/linux-x86_64/ocpirun: ditto

+++ core/dataplane/rdma_driver_interface/interfaces/include/DtTransferInternal.h
@@ -165,7 +165,7 @@ namespace DataTransfer {
     XferServices* getService(EndPoint *s_endpoint, EndPoint *t_endpoint);        
 
     // Create an endpoint for remote hardware
-    EndPoint& allocateProxyEndPoint(const char *epString, uint64_t size);
+    EndPoint& allocateProxyEndPoint(const char *epString, uint32_t size);
 
     static bool canSupport(EndPoint &local_ep, const char *remote_endpoint);
     // Constructors/Destructors
